refactor into separate files for each go file
